% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_mf_app}
\alias{run_mf_app}
\title{Launch the MF Attribution Shiny App}
\usage{
run_mf_app(...)
}
\arguments{
\item{...}{arguments passed to \code{shiny::runApp()} (e.g. port, launch.browser)}
}
\description{
Launch the MF Attribution Shiny App
}
